Remove bucket name initial character tests that gave potentially false positives.
check_good_bucket_name explicitly said it cannot be used to test
the initial character. This is because of cleanup and bucket name
uniqueness limitations.

This leaves a hole in the test coverage: buckets with both letters
and digits as initial characters must work. This needs to be fixed
later.