Tests failing with v4 sigs because of boto bug
There is an unresolved bug in boto in the signature v4 signing code. The
issue is that the ordering of the signed headers is wrong. See
https://github.com/boto/boto/issues/3749
https://github.com/boto/boto/issues/3822

When applying the fix from https://github.com/boto/boto/pull/3824 all
these tests pass.

Tag these tests with aws4_boto_header_ordering_bug so they can be
skipped.