rgw: az: Add archive zone test cases
Test cases for the archive sync module feature. It requires to set up
one configuration with two available zones. One of those zones is the
archive zone (az):

http://docs.ceph.com/docs/nautilus/radosgw/archive-sync-module/

To run the test cases you need to configure 'host_az' and 'port_az' in
your s3test conf (see s3tests.conf.SAMPLE) and set up RGW_S3_USE_AZ=1 in
the environment. Use 'host' and 'port' to configure the non archive
zone.

If RGW_S3_USE_AZ is not available in the environment the test cases will
be skipped. It is the default behaviour.

To run the archive zone test cases directly:

$ RGW_S3_USE_AZ=1 S3TEST_CONF=./s3tests.conf \
./virtualenv/bin/nosetests -v -a 'archive-zone' s3tests_boto3.functional

Signed-off-by: Javier M. Mellid <jmunhoz@igalia.com>