Use a source-built nose instead of one in PyPI
This patch is prompted by the fact that our s3-tests crash
on Fedora 35 with "AttributeError: module 'collections' has
no attribute 'Callable'". This happens because Callable is
in collections.abc in Python 3.10.

Well, nose is unmaintained, so it's bound to break, right?
Wrong. As we found out, the real culprit is actually PyPI.
The nose in Github works, the nose in Fedora works, but
the wheel at PyPI is broken. Someone hacked up the wheel
that pip installs and made a mistake. But with PyPI there's
no recourse.

Our soultion is to switch to nose from the origin in github.

Related-Tracker: https://tracker.ceph.com/issues/53478
Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>