tests: fix test_multipart_resend_first_finishes_last
Original test_multipart_resend_first_finishes_last was doing something very strange
and failed both on our implementation of s3 api and on minio's. Due to to the name and logic,
seems like the resend logic should have been checked. So the test was updated accordingly to
verify that if we resend the same part, but with different contents, the part with the
specified ETag in the CompleteMultipartUpload request is actually applied.

Signed-off-by: Evgeniy Zayats <zayatsevgeniy@nspcc.io>